rngd: allow specification of the amount of entropy in source
The entropy_count was always assumed to be 8-bits for input random data
from rngd. While its impossible to tell without introspection of the source
device how much entropy a data stream has, it seems like some people might
want to be more conservative in that estimate, this change allows a user to
specify the amount of entropy in a given source.

Signed-off-by: Neil Horman <nhorman@tuxdriver.com>
Signed-off-by: Ricardo Neri <ricardo.neri-calderon@linux.intel.com>